Move handlers utils to ecs-agent (#3695)
