+++ agent/handlers/v4/container_association_handler.go
@@ -21,7 +21,7 @@ import (
 	v3 "github.com/aws/amazon-ecs-agent/agent/handlers/v3"
 
 	"github.com/aws/amazon-ecs-agent/agent/engine/dockerstate"
-	"github.com/aws/amazon-ecs-agent/agent/handlers/utils"
+	"github.com/aws/amazon-ecs-agent/ecs-agent/tmds/handlers/utils"
 	"github.com/cihub/seelog"
 )
