+++ ecs-agent/tmds/handlers/utils/helpers.go
